Use an appropriate working directory when launching the Tor Browser.
The Tor Browser assumes that the working directory is the directory
where the browser lives (e.g. the executable). For instance, the
fonts.conf (for fontconfig) the Tor Browser uses looks for the fonts
in the "fonts" sub-directory of the current working directory. This
will fix issues with "empty boxes" as characters in the Save As/Open
dialoges.

Will-fix: #11097