Do the fontconfig dance with all browsers.
I.e. also the Unsafe and I2P Browser's. The main reason is so we do
not need different images for these and the Tor browser in the
automated test suite. One may argue, though, that for the Unsafe
Browser this list of fonts would be fingerprintable, but I would be
surprised if it already isn't so due to the Tor Browser Firefox
patches.