Support newer versions of Vagrant.
... than 1.4.x. Using the Vagrant internals just doesn't seem like a
sustainable approach given how fast things move with
Vagrant. Presumably the commandline interface is more stable.

Newer versions of Vagrant has built-in support for verifying base box
downloads so we can drop our monkey path. This effectively makes us
dependant on such newer versions since Vagrant will fail with an error
if we try to set e.g. config.vm.box_download_checksum when it is not
supported, so we don't have to explicitly add a version check (even
though that might be a good idea).