Update most images that included the Vidalia icon.
For those of them that are meant to show where a small GUI
element (icon) is, I'm providing a new screenshot, that includes the Tor
Status icon, to give more context.

For those that don't really require this much context, I've simply taken
the old image (that was better than the one I could easily come up with
myself) but dropped the Vidalia icon: in this case it's no big deal if
the Tor Status icon is missing, and it may be that we update that icon
soonish (#6841#note-31) so let's avoid unnecessary image churn.

refs: #11128