Build in a simple tmpfs, not a crazy aufs stack.
This is just simpler and safer given the bugginess of aufs, and its
non-standard properties. For instance, there were some concerns that
it doesn't support capabilities (but does tmpfs?) which has some
effect on some Debian packages post-install scripts (e.g. use
capabilities instead of setuid for binaries like ping etc).

Previously the aufs trick allowed su to "share" the on-disk Git repo
and this save that amount of RAM when building in RAM. This was mostly
done to allow Tails to be built on a system with 8 GiB or RAM, but
that is not possible any more, so let's drop this risky optimization.