Let ruby-rjb guess JAVA_HOME instead fixing on one jvm version.
The appropriate jvm to use differs between Debian versions, so setting
a specific one is not convenient when supporting multiple Debian
versions. For instance, the current hardcoded value, which fixes on
the Java 7 jvm, breaks rjb in Debian Sid, so the automated test suite
cannot be run at all.

Let's instead not set it explicitly, and rely on the Debian-specific
(via 0005-Fill-JAVA_HOME-with-a-sensible-value-if-not-set-when.patch)
jvm guessing mechanism.

Will-fix: #11190