Redesign the Dogtail class hierarchy.
The ScriptProxy part was a remnant from the time when I was still
concidering supporting Dogtail's procedural API, but it's now clear
that it will be too hard and not offer anything beyond the Tree
API. Also, the `interact` + block idea was from this time, and also
from a part of rewritten Git history where each use of the Dogtail
class instance inside the? `interact` block built up a single script,
that then was executed. However, since you can mix-in other calls
(e.g. to Sikuli) that may expect that a change already has happened,
that seemed like a bad approach.

So, let's KISS and simply settle on the classes Application and Node,
roughly reflecting those classes in the Tree API. The diff might look
scary, but it just moves code around, and removes some confusing
parts.