Be less smart about starting Chutney.
If we are restoring a snapshots from a previous run (with
--keep-snapshots) where Chutney already has been set up, and where the
network is not plugged, and then we plug the network, then Chutney
will not start at all according to the current logic, because it's
only ensured to be running when restoring from a snapshot where the
network already was plugged.

So, let's just ensure Chutney is running before each @product feature,
even if it won't be used at all.