Test bridges and pluggable transports with Chutney.
We now set up all bridges (from now on that term also includes PTs)
using Chutney, so we no longer need to provide them in our "secret"
local configuration any more, which is a big plus in itself.

This includes replacing one of the authorities with a bridge
authority.

Note: the firewall test (i.e. that only the expected hosts are
contacted) is pretty broken at the moment, definitely in the bridge
mode case, but perhaps in normal operation too, since what we check is
which hosts are contacted, and the complete Tor network *and* bridges
now run on the same host (and IP address).