Add static sleep to deal with tcpdump/pcap processing latency.
One instance where I've seen this is in the "Connecting to an SSH
server on the LAN" scenario. The last step ("I am prompted to verify
the SSH fingerprint for the SSH server") can finish so quickly, that
when we kill the @check_tor_leaks sniffer it hasn't processed the
traffic between Tails' SSH client and the SSH server on the LAN yet.