Reorder the page to be more upfront
Give the screenshot first as it clarifies the context and makes it
easier to understand what the elements that we are describing. Then
explain the different elements and explicit the example on the
screenshot.

The diagram of a Tor connection might still be useful to understand how
this relates to what's going on on Internet, but we can put it at the
end.