Make Electrum tests pass after bump to 2.6.3-1.
The new version decreases the size of the seed text box (the old image
expects a larger one), and it is not focused by default. Furthermore,
there seems to be a race between the where Sikuli could detect the old
text box after clicking next, so let's properly wait for the
verification window.