Drop the hard requirement of OpenPGP communication for mirror operators.
This change doesn't imply any regression, compared to the current state
of things, given we've already somewhat given up on this in practices
for a long time (the work required to maintain the corresponding OpenPGP
keyring was too much compared to the perceived advantage).

See the "[RFC] Dropping requirement for OpenPGP communication with HTTP
mirror operators?" thread on tails-dev@, for details.