Always run the 'the Tails desktop is ready' step.
... after logging in via Tails Greeter. Now we do some very important
stuff there, like enabling the accessibility toolkit in GNOME. That is
specifically something we lose if we forgot to run that step after
rebooting, which specifically was a problem in at least one scenario
(Deleting a Tails persistent partition).

While we're at it, let's also run the 'Tails Greeter has dealt with
the sudo password' step at the same time.