apt-get wrapper: deal with an unlikely corner case.
We know that apt-get's action (install, update, etc.) will come _before_ any
package name, so as soon as we see a supported action, we stop looking for one.

The only problem left is if apt-get is called with an unsupported action,
followed by a package name equal to one of our supported actions. If this were
to happen, even though unlikely, then mode=$package and likely the resulting
apt-get command-line should be invalid or break the build sooner or later, so
I'm not too concerned about this breaking stuff without us noticing immediately.