Drop the usage of Tor Check in our tests.
It doesn't make sense now when we use Chutney since that always means
it will report that Tor is not being used.

In some scenarios we used it simply because we needed *some* page
(distinct from the Tails News page) and we had the image. In those
cases we now use the Tails homepage instead. I tried to use dogtail,
but realized that dogtail won't work for the Unsafe Browser due to