APT pinning: with our snapshots system, backports have o=Debian.
... and here as well, add a sanity check to warn the developer early if
they used "o=Debian Backports".

We could easily fix that for time-based snapshots, but for tagged
snapshots it's a different story: when we pull packages from time-based
snapshots into tagged snapshots, we have no means to know what was
the Origin they were pulled from initially.

Granted, we *could* handle backports as a different archive (just like
we have to do for debian-security), but it doesn't feel worth
the hassle.