Add explicit APT pinning for libdvd-pkg.
So far, it was magically pulled from jessie-backports merely because we
forgot to add a general pinning level of -10 for backports in
config/chroot_apt/preferences, unlike what we've done for all our other
APT sources. And now it's broken, because in our APT snapshots system,
backports have o=Debian, and not o=Debian Backports anymore.

So, let's remove this exception, and pin the package explicitly.