Run tcpdump with --immediate-mode.
With this option, "packets are delivered to tcpdump as soon as they
arrive, rather than being buffered for efficiency". It's not available
in Jessie's tcpdump, but it is in jessie-backports.

Background: Normally tcpdump uses a 1000 ms = 1 s timeout (set via
`pcap_set_timeout`; for more details, see the "read timeout" section
in the "pcap" man page from the libpcap*-dev package) which explaines
that the `sleep 1` worked most of the time before on my lightly loaded
system, but much rarely on jenkins highly loaded system.

At the moment this should hopefully fix the issues we have with the
"Anti test: Detecting UDP leaks of DNS lookups with the firewall leak
detector" scenario in the test/9521-chutney branch.