Remove obsolete doc (Will-fix: #11339)
This section was added as a workaround since it was not possible to
copy and paste into Pinentry. This is now possible again as of #11239 so
let's remove this section.

Being able to copy and paste is the expected behavior and shouldn't need
to be explicitly mentioned here (especially since we don't have a
dedicated page for Pinentry).