Add basis for toggleable dollar/euro donations
What I want to do here is not possible using ikiwiki/toggle.js only, so
I'm doing a quick JS prototype with Bootstrap and jQuery. We might later
on prefer rewriting only what's need in vanilla JS or dropping the idea
all the way.

I'm also not including the third-party scripts in Git on purpose not to
unwanted data; but the URLs are in the code.