Release process: don't freeze APT snapshots for point-releases.
For the stable branch, generally we want to use snapshots labeled with
the Git tag of the latest Tails release. If we want to ever upgrade
stable to a fresher set of snapshots, we'll encode them in
config/APT_snapshots.d/*/serial explicitly, but that's an exception (and
is managed in Git, so can go through our usual review'n'merge process),
not the default case.