Fix multiple issues with dogtail waiting for a browser page to load.
Also, this is a great simplification. :)

It turns out the previous code had multiple errors (symbol error
because of True when it should be true, no .wait so the generated
python code was never run) but even after fixing them I learned that
try_for and assert_raise doesn't work together because it will catch
try_for's timeout exception. Even when telling assert_raise to look
for a specific exception it will still catch the other ones and then
interpret them as an assertion failure, so the try_for never
halts. Yay. It should be possible to get an ~assert_raise that we need
here with some begin-rescue logic, but let's not bother.