Test suite: stop running "the Tails desktop is ready" twice during some scenarios.
commit:fc777723 dealt with duplicates calls to "the Tails desktop is ready",
except it missed the fact that "Tails seems to have booted normally" was an
alias for it. So, here we remove all such duplicates, and in the end we remove
"Tails seems to have booted normally" as well since it's not used anymore.