Hide dogtail's AT-SPI Browser in the Applications menu.
Apparently some users are curious enough to browse the Programming section,
start "AT-SPI Browser" from there without knowing what it is about (let's call
this "exploratory testing" :) and as a result they are confused when they see…
an AT-SPI Browser appear.

Regardless of what we may think wrt. how we should generally handle such
scenarios, let's think in an ad-hoc way here: this development helper is not
meant to be useful to end-users, so let's hide it to avoid such confusion;
developers who need AT-SPI Browser will know how to start it.

refs: #11498