Add memory_wipe to the prereqs of all init-top initramfs scripts that are not ours.
This is the only way to have guarantees as to the order in which those
scripts are executed. The practical effect is that when erasing memory,
our memory_wipe script will be run first, and since it shuts down the
system in the end, none of the other init-top scripts will be executed.

refs: #10733