Optimize scenario by not using the network.
Here we don't exercise any functionality that is affected by
networking, so let's kill the delay (waiting for Tor to bootstrap)
that enabling networking introduces.

For general performance, but also since we just removed the @fragile
tag from this scenario without a review'n'merge (or much testing);
let's at least avoid introducing a dependency on (historically unreliable) Tor
bootstrapping to Jenkins.