You cannot `return` in a step definition block.
It will throw a JumpError because a block is not a function, and hence
you cannot return from it. For a block the corresponding thing would
be `next`. The lack of JumpError's in Jenkins suggests we never hit
this case, actually. While this code is more readable, I wonder
whether it still fixed the TOCTOU it was meant to fix -- we won't get
a FindFailed, but we can have sikuli find the target, and while it is
moving the cursor to click it, it disappears, so the click is done on
nothing and the notification applet hasn't opened => we fail ten
seconds later. However, that hasn't been observed on Jenkins; and this
was introduced as part of the fix for #11464, and the rest clearly has
value, so let's see if it will work any way.