Reduce memory wipe timeout by a factor of 10.
The previous crazy value of 5 minutes per gigabyte was when we were
trying the automated in a nested virtualized environment with crappy
hardware for that. And even then it was a big exaggeration! In our
current setups I want to know if the wipe takes more than 30 seconds
per gigabyte -- that would imply a bug.