Revert "Give the VM used for the test suite a hardware RNG device."
This reverts commit a47df202ed091ff64c73048e3879c5a7a9797f4c.

Until libvirt 1.3.4, once can only use /dev/random (which is blocking)
as a backend for virtio-rng, and as a result I see weird test suite
failures in which the system under test freezes.

Once we can use libvirt 1.3.4+ and want to actually implement tests
that exercise rngd, we can revisit.

refs: #5650