Remove unecessary ESC+TAB dance in the boot menu.
The new step is supposed to stop at the kernel command line prompt, so
there's no real reason to hide it again for the next step. Let just type
directly the options we want. Also it may be a bit racy: I've seen
scenarios where the ESC key was typed and the kernel command line
hidden, but then shown again quickly after, so Sikuli couldn't find the
TailsBootSplashTabMsg image.

Refs: #10777