Stop shipping ttdnsd.
It was only useful for developers and power-users who can install it themselves
as needed. It's been unmaintained upstream for many years. It's very buggy so we
had to remove it from the DNS resolution loop years ago. It's not in Debian.
And it's one of the only two bits of Tails that still rely on tsocks, that is
RC-buggy, unmaintained in Debian, and not in Stretch at the moment.

So it has become clear that the cost of keeping ttdnsd now outweighs the
benefits it brings.

refs: #10959