Use torsocks to torify Git, and drop tsocks entirely.
tsocks has been unmaintained for years in Debian, and is not in testing
at the moment.

Since we dropped ttdnsd, the only use case for tsocks was torifying Git.
But thanks to torsocks 2.1's AllowOutboundLocalhost we can now use torsocks to
torify Git without breaking Git-over-SSH. See commit:3922b0b2 for the previous
iteration of this.

refs: #10959
closes: #10955