Revert to GNOME's default font antialiasing/hinting.
We fixed on rgba/slight when converting some manual fontconfig stuff
to GNOME's layer on top of it, but at least from a fresh Stretch
install (2016-08-24) I get grayscale/slight. It could be that some
auto-detection is involved, so the values would be different depending
on the actual hardware. Any way, let's try to decrease our delta here.