Drop patch to keep haveged, saned, spice-vdagent and laptop-mode running on shutdown.
These patches are no-ops on Stretch, where these services have native
systemd unit files.

IMO it's not worth porting these patches: saned is socket-activated so
in most cases it does not have to be shut down, and I expect that the
other ones can be stopped pretty quickly.

Let's bring back this kind of optimization if, and only if, we identify
an actual problem to fix in this area :)