Never (pretend to) thaw APT snapshots on the stable branch.
Let's always encode, on the stable branch, the exact set of APT snapshots we
want to use in next point-release. Previously we would pretend to thaw them by
writing "latest" in APT_snapshots.d/*/serial, but then apt-mirror would
special-case this situation: on an unreleased branch based on stable, it would
consider that "latest" means "stick to previous release's tagged snapshot".
Which worked fine at build time, except the part of our build system that
creates the build manifest does not know about this convention, so the resulting
build manifest would point to the latest APT snapshots we have, even though they
were not really used during the build. And even if the build manifest pointed to
the right place (i.e. the previous release's tagged APT snapshot), which is the
only place where some of the needed packages are when tagging the next
point-release's APT snapshot: tails-prepare-tagged-apt-snapshot-import does not
know how to generate a configuration that pulls from there.

So let's drop this special meaning of "latest", and make things simpler by
actually hard-coding in Git the snapshots we really want to use.

This implies the documentation change that makes sure that we're keeping these
time-based APT snapshots long enough.

refs: #11612