Test suite: update the screenshot scenario.
The Pictures & friends directories are not automatically created on
feature/stretch currently, and thus gnome-screenshot saves screenshots to $HOME.

Given these directories would not be persistent, IMO it's a feature to avoid
encouraging users to store stuff in there. So I'm simply adjusting the test to
reflect how it now works.