Hide OpenPGP Applet launcher from the Applications menu.
Until we have something (else than the systray icon) to launch, i.e. other bits
of #7433 are done, it makes no sense to have a launcher there in the context of
Tails, since we launch the applet by default, and added a confirmation dialog
for exiting it. And keeping the launcher around means that users will try
to run it more than once, which is not very well supported.

Note that we have to *move* the desktop file to /etc/xdg/autostart/:
both Hidden=true and OnlyShowIn= would not only hide the launcher,
but they would also inhibit the automatic startup of the applet.

refs: #11781