Remove Icedove scenarios using hidden services.
A fundamental limitation of using our simulated Tor network (via
Chutney) is that We cannot access the hidden services of the real Tor
network. A possible future solution is that we tell Chutney to start a
hidden service pointing to the real host of the Email provider. Or that
we just run the SMTP/POP/IMAP servers ourselves.

Even without Chutney, the approach currently chosen is problematic when
using hidden services which use the SSL certificate for the clearnet
service (like Riseup) -- in the autoconfiguration wizard it just results
in the password test failing, but if we do it through an Advanced
setup (which puts us in the normal account configuration interface) then
we get a prompt where we can tell Icedove to accept the cert any
way. This inconsistency is a bug in Icedove, of course, but it shows
that the above solution of starting "a hidden service pointing to the
real host of the Email provider" will fail if SSL is mandatory (like
with Riseup).