Make it possible to start the automated test suite with --keep-snapshots after clearing the tmpdir.
Now you can, for example:

    ./run_test_suite --keep-snapshots ...
    # Realize that you want to run with a different ISO, so snapshots
    # must be cleared:
    rm -r path/to/tmpdir
    ./run_test_suite --keep-snapshots ...

Previously the last step would fail since libvirt still is aware of the
Storage Pool, so it would try to restore it but error out since it
doesn't exist. So you would have to work around it with something like:

    ./run_test_suite ...
    # Wait until the VM starts booting, then <C-c>
    ./run_test_suite --keep-snapshots ...

Note that the only changed code path is if --keep-snapshots is given, so
this commit will not affect Jenkins, only expert developers that use
--keep-snapshots => reviewers don't have to be scared of this commit. :)
Also, the boolean condition (DeMorgan!!!) can be simplified, but I
thought it was clearer to keep it "expanded" like this.