Install Linux 4.7.
While preparing Tails 2.6 it was discovered that our previous last
minute bump of the 'debian' snapshot (to get Jessie 8.6) caused
linux-compiler-gcc-4.9-x86 to be updated to a version from another Linux
source version (4.7) than what we planned to install (4.6), resulting in
the tagged snapshot being impossible to create since it would require
two different linux source packages which we do not support.

Note that this includes bumping the aufs4-standalone Git submodule to
the latest branch for Linux 4.7.