Port tails-security-check to use Tails::Download::HTTPS.
This is needed to pin Let's Encrypt _intermediate_ CA, which is not supported by
IO::Socket::SSL and Net::SSLeay. For details, see:
https://labs.riseup.net/code/issues/11810#note-1.

Besides, it's much cleaner and easier to maintain to have one single (somewhat
hardened) HTTPS download implementation, that's covered quite a bit by
tails-iuk's test suite, instead of having to maintain a second one just for
tails-security-check.

refs: #11810