Make the tor-controlport-filter more oblivious about the control language.
The way we used stem actually made us look quite deep into the language
-- doing some semi-rigid parsing, and then try to find the exact
equivalent stem-function for the processed commands. That's quite
unnecessary when all we need to be is a "blind" proxy between the client
and control port. However, for async stuff like events that is harder,
so let's leverage stem there.