Make tor-controlport-filter run as root with full capabilities.
It requires to lookup which PID the client have (via the connecting
port) which normal users cannot, and no matter how I tried to set
CapabilityBoundingSet, it didn't work. OTOH, the attack surface of this
script should be pretty small so running it as root is perhaps not so
bad.