tor-controlport-filter: also match filter based on the client's user.
Sometimes the executable path just isn't enough. For instance, for the
tor-launcher filter the executable is the unconfined firefox executable,
also used by our chroot browsers. So let's be a bit more restrictive.

While we're at it, make it possible for a single client to match
multiple filters -- otherwise the rules for which single filter will be
selected in case multiple matches will just complicate things. So, if we
want, we can now refactor common parts of filters. :)