tor-controlport-filter: add option restrict-stream-events.
When set, any STREAM events sent to the client (after it has subscribed
to them) will be restricted to those originating from the client itself.

For instance, if set for the Tor Browser this means that it will only
receive its own STREAM events, and not those from other applications,
which are not used for its per-tab circuit view any way.