tor-controlport-filter: non-atomic commit, sorry!
Most of it is about ending the stupid experiment with multiple filters
and merging, so this is a huge simplification made possible by dropping
some functionality of questionable use. One filter to rule them all!

It also adds the `confs` syntactic sugar, and fixes some tiny issues
introduced in recent commits.