Don't remove the whole CDROM device when we want to eject it.
The remove_cdrom code now unplug the whole device, which does not fit
when we just want to eject it. In place we use the `eject` command
inside the VM. It may be a bit less realistic than previously, but with
QEMU 2.6 we can't set the source file to an empty string.

Will-fix: #11874