Avoid unnecessary delay in the remote shell.
See the added code comment. BTW, thanks to the defensive (foreboding?)
approach taken by commit 5020062, when we switch from sigwait() to
sigtimedwait(), we avoid this being a deadlock.

Coincidentally, this code is more elegant, with a clearer flow of
execution and no nested loops, and it reminds us to send a mourning
thought to those lost signals out there. ${Deity}speed, brave electrons!