Skip Dogtail cleanup due to remote shell instability.
If the remote shell times out while running the dogtail script, this
cleanup will try to use the shell when it is dead, resulting in another
call that will time out. Since nothing is written to stdout, this may
result in Jenkins aborting the whole run due to this extra timeout.

Besides, we don't really care if potentially hundreds of scripts are
littering the guest's /tmp.