Let's only force new releases of bundled .deb:s for each major release.
Rationale: the translation work is not moving very fast these days,
being quite polished for many languages, so we gain very little from
frequently pulling it and preparing releases *only* for new translations
(no new code). So, let's relieve the RM from this work for most
releases, but still have a process in place for making sure that
translations are pulled regularly.