Remove dpkg's .md5sums files.
These files are pretty useless for us and mainly occupy space on the
image. They are, for instance, not useful for checking the authenticity
of the filesystem (an external verification tool and source of these
checksums would be required), and checking for corruption is less
relevant in Tails' context, where the system partition is read-only (the
point being: if they do differ, chances are problems would manifest in
much more obvious ways).