Don't ship a /etc/machine-id for increased reproducability.
Also, let's consider leaks/fingerprintability: if we ship this file, it
will be identical for all users of the same Tails build, but if it's
generated per-boot, Tails sessions can be uniquely identified by
it. It's hard to way these against each other, so I don't think this
aspect blocks the removal. However, we could ship a static one in *all*
Tails builds without losing reproducability. Does that gain us anything?