Move intltool-update --report statistics to language_statistics.sh
We deleted the intltool-update --report from refresh_translations.
See #11967.

This happened because intltool-update --report does not only report
but modifies all the PO file headers which it should not do.

Here we created an ugly workaround so that only the report is made, and
modified files will be replaced by their original files afterwards.