Force upgrading with debian-security APT repo activated.
Otherwise each generated basebox will already be behind some
security upgrades when it is created, and the creation of the vagrant
builder based on this basebox is delayed by applying this security
upgrades. Let save a bit of time on VM creation by including the latest
security upgrade when creating a basebox.

Will-fix: #11980