Ship all build dependencies in the basebox.
They take quite some time to install every time a new vagrant builder is
provisioned. Shiping them reduce the time it takes to build an ISO, and
grow the final basebox image of around 30MB. It's probably fastest to
download this in the basebox than installing them each time.

Will-fix: #11980