Dogtail: use the remote shell's new Python session feature.
... to significantly improve Dogtail's performance by saving state and
reusing it between Dogtail commands.

This is a massive commit, and it changes the semantics of the creation
of Dogtail objects. Previously they just created the code that then
would be run once an actionable method was called (.wait, .click etc),
but now it works like in Python, that Dogtail will try to find the
graphical element upon object creation.

Will-fix: #12059