Lower systemd's DefaultTimeoutStopSec (refs: #12061).
session-c1.scope won't shut down before the default timeout (1min 30s),
and I didn't manage to change its timeout, nor to let shutdown.target
ignore it.

So let's change the default timeout globally. This feels hackish, but OTOH it
means that shutdown won't ever be blocked by any buggy service that takes
a while to stop, which can be seen as a safety feature in the context of Tails,
since we want to wipe memory afterwards. Maybe.