Website: explicitly add sort="age" to all inline directives I can easily & mechanically patch (refs: #12053).
A serious performance regression was brought by ikiwiki 3.20150614. One of the
few changes introduced in that release was to change the default sorting, for
inline directives, from "age" to "age title". Presumably the latter is
substantially slower, possibly since it has to process the *content* of all
included pages in order to retrieve their title, which may mean that a subset of
pages are processed more than once.