Undo some refactoring and remove the CommandFiltered exception.
Those changes (which include several errors) made it harder to deal
with control flow -- the current code is pretty much based on that we
can `continue` and `break`, so I've restored the loop.

So, this reverts 7b5683e21206d61d0a360db8edc6610832428301 and replaces
it with a different solution (fixing the same issue as that commit).