CSS: drop forced height that makes the infography hide the text on narrow screens.
On narrow screens we display the infography on top of the corresponding text,
while on wider screens we display it on its left.

Without this change, on a narrow (e.g. 720px wide) screen, whenever the
infography of the IA has 2 images or more in a single .step-image div, that div
is forced to a 200px height (but really the 2 images use 2*200px height), so the
text that follows is displayed with a 200px offset, i.e. part of it is hidden by
the second image. On a smartphone with an Android browser I was not able to make
that text visible.

So, let's not pretend all .step-image div:s are 200px high (they are not), and
instead let the browser calculate how much screen estate they need.