Ensure the VirtualBox guest DKMS modules are built for the kernel we want them for (refs: #12139).
In some situations, depending on the version of the running kernel, the modules
would not be built for the 686 kernel, which is the one that needs the
VirtualBox guest modules.

This commit ensures the VirtualBox guest modules are built and installed
regardless of the how the build environment looks like.