Disable -proposed-updates at boot time.
If a Debian point release happens right after a freeze but we have
decided to enable it before the freeze to get (at least most of) it,
then we get in the situation where -proposed-updates is enabled in the
final release, which we don't want. We only want it enabled at build
time.

Will-fix: #12169