Upgrader: use the alpha channel when the next version will be an alpha, beta, or RC.
The basic idea here is to make it easy for users of 3.0~betaN to upgrade to the
next beta or RC, without having to type any command-line, and without having to
create UDFs for these upgrade paths on the "stable" update channel (that would
feel wrong, wouldn't it?).

Bonus: once this is in place, more generally we'll be able to update our release
process doc+scripts to stop creating meaningless "stable" channel upgrades for
RCs, such as upgrade/v1/Tails/2.10~rc1/i386/stable/upgrades.yml, that we have
been needing so far since our RCs thought they were using the "stable"
update channel.

refs: #12206