Test suite: let's see if the 'press' action is more reliable than .click().
The .click() method actually looks up the coordinates of the element
and then clicks those coordinates. The 'press' action directly causes
a "click" event that cannot be lost.

The affected part has been failing quite regularly on Jenkins,
presumably because the GUI isn't finished loading at the time, so the
coordinates we end up clicking on (with .click()) are the wrong ones.