Don't include /var/lib/initramfs-tools/*-amd64 in the SquashFS.
Due to #12330 (which is about the ISO filesystem, really), this file indirectly
makes the SquashFS unreproducible as it contains a hash of the initrd.
I'm pretty sure this file is useless in a live system, so let's drop it for now,
so that I can tackle #12032 (by giving mksquashfs stable inputs) earlier.