Vagrant: make the checked out source (not only .git) available to the builder.
This branch already assumes it to be the case in some places, and it's
handy since with an actual checkout, we can use scripts, like
auto/scripts/apt-snapshots-serials, that we need. We also adapt the
build-tails script to this change.

(Note: This change is actually already present in another
branch, wip/11972-use-vagrant-in-jenkins branch, but it's not suitable
for merging into this one yet.)