Vagrant: provision a new Vagrant VM for every ISO build.
Well, at least when there are only successes; on failure the VM is
kept and a new `rake build` will start an immediate build, reusing the
cached wiki from last time. This should be a nice trade-off between
making it fast to rebuild a branch after a failure, and maintaining
the "staticness" and "up-to-dateness" of the build environment.

Will-fix: #11980, #11981