Vagrant: avoid scary scp warnings.
If you use the `rake vm:*` commands creatively you may end up in a
state where the SSH fingerprint has changed, showing a big warning
each artifact we fetch that makes me think "shit, the build failed!"
when it in fact succeeded and artifacts were fetched just fine.

Since we don't care about that verification mechanism we might as well
disable it completely (like this commit does) to get rid of any future
possibility for this to become an issue again.